There was a problem with realm configurations with specified schema version.
Realm browser plugin could not open a realm instance if no schema
version was provided.

As we dont know schema version from the realm files, in order to not
request that realm configuration to library users, we can change way
Realm is used.

Instead of usign Realm in the normal way, we use DynamicRealm
https://realm.io/docs/java/latest/api/io/realm/DynamicRealm.html

"Dynamic Realms do not enforce schemas or schema versions and
RealmMigration code is not used even if it has been defined in the
RealmConfiguration."